requirements.txt
streamlit
confluent-kafka
boto3
python-dotenv
pyarrow
pandas

